requirements.txt
environs==7.3.1
aws-cdk.core==1.79.0
aws-cdk.aws-ec2==1.79.0
